Add new surface figures
